Create CONTRIBUTING.md
As https://opensource.guide/ it is best practices to have a CONTRIBUTING.md file, I got the one that we have on the website and I have converted it to .md. I hope it is helpful. :)
If you also click here https://github.com/apache/hop/community you will see that github is pointing out this.