Update http.adoc
https://hop.apache.org/manual/latest/pipeline/transforms/http.html#_description

When viewing in a browser, on the documentation page for the HTTP Client, the text in the rows for the columns Option and Description begin to get mixed up at "Connection Close wait time", which results in a a mishmash where Option text is in Description and they don't align correctly moving down the table. For example, the last row displays as Option: Area where you define the parameter name-value pairs to pass on the URL | Description: Custom HTTP Headers.

It seems there was a missing "|" (vertical bar) on line 45 after "Connection Close wait time". By adding the vertical bar there, the rest of the values in the table will be aligned in their correct columns.