Create transform_dialog_standards
To get us going, I wanted documentation for people to hit to understand what fields show up in various dialogs and their purpose.
In the future we should clean this up and have steps only present what they can truly alter.  If there is missing functionality, we can have it added to the transforms.